version 1.14.4-0c it works and everything. i just don't think it's publishable. for a concept that sounds so polished, i think people will be disappointed about the limitation with subspaces based on enchantments and with the portal linking order.
- craft the gateway prism normally
- to link it to a subspace, choose
up to eight dyes to place around it
in the crafting grid-- NO; i know what
to do now!
- to link it to a subspace, combine it
with a banner in the anvil-- or, something

- there is still the idea of clicking the
first portal with the gateway prism, then
clicking the second portal. when the second
portal is clicked, the second portal will
activate and the gateway prism will belong
to it. clicking on the first portal saves
that world,x,y,z, transferring it to the
second one. I like this idea, even though
it is a bit more work to make it behave
like the one i'm doing right now.

- perhaps... make two more items?
a gateway prism is used for the above idea.
a gateway beacon is used to make the teleportal
accept teleports via a specific key- the name
of the gateway beacon.

i'll think about it more.